test: add test for extra fields
